--- CS320_ResistorSorter/bin/.gitignore
@@ -1,1 +1,0 @@
-/resistorSorter/
