+++ CS320_ResistorSorter/old files/src/resistorSorter/Bin.java
